Add functionality to create new issue
A new function, `create_issue`, has been added to allow users to create new issues. This function adds a new issue to an array of existing issues. Additionally, minor adjustments have been made to the UI to accommodate this feature, as well as changes in some of the existing functions and messages.